e2e -> test/e2e
