+++ test/e2e/httpproxy/008_tcproute_https_termination_test.go
@@ -16,47 +16,45 @@
 package httpproxy
 
 import (
+	"context"
+	"crypto/tls"
+	"crypto/x509"
 	"testing"
 
 	contourv1 "github.com/projectcontour/contour/apis/projectcontour/v1"
-	"github.com/projectcontour/contour/e2e"
-	"github.com/stretchr/testify/assert"
+	"github.com/projectcontour/contour/test/e2e"
 	"github.com/stretchr/testify/require"
+	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-func testHostHeaderRewrite(t *testing.T, fx *e2e.Framework) {
-	namespace := "017-host-header-rewrite"
+func testTCPRouteHTTPSTermination(t *testing.T, fx *e2e.Framework) {
+	namespace := "008-tcp-route-https-termination"
 
 	fx.CreateNamespace(namespace)
 	defer fx.DeleteNamespace(namespace)
 
 	fx.Fixtures.Echo.Create(namespace, "ingress-conformance-echo")
+	fx.CreateSelfSignedCert(namespace, "echo-cert", "echo-cert", "tcp-route-https-termination.projectcontour.io")
 
 	p := &contourv1.HTTPProxy{
 		ObjectMeta: metav1.ObjectMeta{
 			Namespace: namespace,
-			Name:      "host-header-rewrite",
+			Name:      "echo-tcpproxy",
 		},
 		Spec: contourv1.HTTPProxySpec{
 			VirtualHost: &contourv1.VirtualHost{
-				Fqdn: "hostheaderrewrite.projectcontour.io",
+				Fqdn: "tcp-route-https-termination.projectcontour.io",
+				TLS: &contourv1.TLS{
+					SecretName: "echo-cert",
+				},
 			},
-			Routes: []contourv1.Route{
-				{
-					Services: []contourv1.Service{
-						{
-							Name: "ingress-conformance-echo",
-							Port: 80,
-						},
-					},
-					RequestHeadersPolicy: &contourv1.HeadersPolicy{
-						Set: []contourv1.HeaderValue{
-							{
-								Name:  "Host",
-								Value: "rewritten.com",
-							},
-						},
+			TCPProxy: &contourv1.TCPProxy{
+				Services: []contourv1.Service{
+					{
+						Name: "ingress-conformance-echo",
+						Port: 80,
 					},
 				},
 			},
@@ -64,11 +62,22 @@ func testHostHeaderRewrite(t *testing.T, fx *e2e.Framework) {
 	}
 	fx.CreateHTTPProxyAndWaitFor(p, httpProxyValid)
 
-	res, ok := fx.HTTP.RequestUntil(&e2e.HTTPRequestOpts{
-		Host:      p.Spec.VirtualHost.Fqdn,
+	certSecret := &corev1.Secret{}
+	key := client.ObjectKey{Namespace: namespace, Name: "echo-cert"}
+	require.NoError(t, fx.Client.Get(context.TODO(), key, certSecret))
+
+	_, ok := fx.HTTP.SecureRequestUntil(&e2e.HTTPSRequestOpts{
+		Host: p.Spec.VirtualHost.Fqdn,
+		TLSConfigOpts: []func(*tls.Config){
+			func(c *tls.Config) {
+				certPool := x509.NewCertPool()
+				certPool.AppendCertsFromPEM(certSecret.Data["ca.crt"])
+
+				c.RootCAs = certPool
+				c.InsecureSkipVerify = false
+			},
+		},
 		Condition: e2e.HasStatusCode(200),
 	})
-	require.True(t, ok, "did not get 200 response")
-
-	assert.Equal(t, "rewritten.com", fx.GetEchoResponseBody(res.Body).Host)
+	require.Truef(t, ok, "did not receive 200 response")
 }
